+++ frontend/src/component/integrations/IntegrationEvents/IntegrationEventsDetails/IntegrationEventsDetails.tsx
