+++ frontend/src/component/integrations/IntegrationEvents/IntegrationEventsDetails/IntegrationEventsDetailsAccordion.tsx
