+++ frontend/src/component/integrations/IntegrationList/IntegrationCard/IntegrationCardMenu/IntegrationCardMenu.tsx
@@ -29,7 +29,7 @@ import { ConditionallyRender } from 'component/common/ConditionallyRender/Condit
 import { useUiFlag } from 'hooks/useUiFlag';
 import Visibility from '@mui/icons-material/Visibility';
 import { PermissionHOC } from 'component/common/PermissionHOC/PermissionHOC';
-import { IntegrationEventsModal } from 'component/integrations/IntegrationEventsModal/IntegrationEventsModal';
+import { IntegrationEventsModal } from 'component/integrations/IntegrationEvents/IntegrationEventsModal';
 
 interface IIntegrationCardMenuProps {
     addon: AddonSchema;
